feat: openapi와 도메인 유즈케이스 및 프레젠테이션 병합
