+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/application/nightstudy/data/NightStudyMapper.kt
@@ -1,7 +1,10 @@
 package com.b1nd.dodamdodam.nightstudy.application.nightstudy.data
 
+import com.b1nd.dodamdodam.grpc.user.UserResponse
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.ApplyPersonalNightStudyRequest
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.ApplyProjectNightStudyRequest
+import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.OpenApiNightStudyResponse
+import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.OpenApiUserInfoResponse
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.PersonalNightStudyResponse
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.ProjectNightStudyResponse
 import com.b1nd.dodamdodam.nightstudy.domain.nightstudy.entity.NightStudyEntity
@@ -54,3 +57,44 @@ fun NightStudyEntity.toProjectNightStudyResponse() = ProjectNightStudyResponse(
 )
 
 fun List<NightStudyEntity>.toProjectNightStudyListResponse() = map { it.toProjectNightStudyResponse() }
+
+fun UserResponse.toOpenApiUserInfoResponse() = OpenApiUserInfoResponse(
+    publicId = UUID.fromString(publicId),
+    username = username,
+    name = name,
+    phone = if (hasPhone()) phone else null,
+    profileImage = if (hasProfileImage()) profileImage else null,
+    status = status,
+    roles = rolesList,
+    student = if (hasStudent()) {
+        OpenApiUserInfoResponse.StudentInfo(
+            grade = student.grade,
+            room = student.room,
+            number = student.number,
+        )
+    } else null,
+    teacher = if (hasTeacher()) {
+        OpenApiUserInfoResponse.TeacherInfo(
+            position = teacher.position,
+        )
+    } else null,
+)
+
+fun NightStudyEntity.toOpenApiNightStudyResponse(
+    leader: OpenApiUserInfoResponse,
+    members: List<OpenApiUserInfoResponse>
+) = OpenApiNightStudyResponse(
+    id = id!!,
+    name = name,
+    description = description,
+    period = period,
+    startAt = startAt,
+    endAt = endAt,
+    needPhone = needPhone,
+    needPhoneReason = needPhoneReason,
+    status = status,
+    leader = leader,
+    members = members,
+    rejectionReason = rejectionReason,
+    type = type,
+)
