+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/application/nightstudy/data/request/RejectNightStudyRequest.kt
@@ -1,0 +1,5 @@
+package com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request
+
+data class RejectNightStudyRequest(
+    val rejectionReason: String
+)
