+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/application/nightstudy/data/response/OpenApiUserInfoResponse.kt
@@ -1,4 +1,4 @@
-package com.b1nd.dodamdodam.nightstudy.application.openapi.data.response
+package com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response
 
 import java.util.UUID
