--- services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/application/openapi/data/request/RejectNightStudyRequest.kt
@@ -1,5 +1,0 @@
-package com.b1nd.dodamdodam.nightstudy.application.openapi.data.request
-
-data class RejectNightStudyRequest(
-    val rejectionReason: String
-)
